Trim for size
Starting with farkos-D/2 the KSM files have to fit in limited space,
so strip out extra print and farkos:ev calls.

setstage has additional changes: now we pass in e and f,
rather than passing in the setstage lexicon; and we return
the name of the auto_stage file, rather than having it hard
coded in several places.