add BIST package
Support construction of Built-In Self-Tests.

Note that as we can not include code in modules that is excluded
from the KSM file, these "Built-In" tests have to be stored
somewhere else in the archive.

The exception is that BIST itself has a built-in test module.

Entry points: logline, result, pass, fail, t, f, eval, eq, ne.