add PHASES
create flight phases that work with the mission
package. provide a _get(n) that fetches the
delegate for the name, returning a DUMMY
if it is not found.

add a comment to pof indicating that it was used
for an early mission. it can be kept around but it
is better to use mission and phases.