fix pause step
the pause flight phase method was waiting internally
rather than deciding what to do and returning a delay.

Improve it to make decisions and return a delay
based on current state, rather than doing a wait.

When waiting for the flight engineer to activate RCS
to continue, keep showing the HUD message.