Unnaturally Belligerent Knot Fixes #24
Can't use AG1/AG2/AG3 but it works to set up
RCS and Lights: both on is 20 m/s, just RCS
is 5 m/s, an RCS OFF is stop.

Jet engines need to be 10% or they flame out,
and I think they can restart but why take
the chance.

Turns out that at 10% throttle on two jet engines
I sometimes creep forward but this was good
enough to pick up a bunch of science.