Meaningfully Holistic Pickle: GO script [wip]
Ascend to a chosen READY orbit that is above or
below the rescue target. Match inclination.

Progress on code to transfer to the target orbit
when doing so matches the target vessel; adds
a maneuver node and tunes it to match by
hill climbing the approach distance as a function
of the thrust vector in the node.