Rendezvous: debug for "target is higher"
Math was good for descending to an orbit that is lower
but not much lower. Did not work for large differences
in velocity, and did not handle the cases where we needed
to burn retrograde.

So rework to handle rendezvous with a target that is in
an orbit far far above us.