mnv_step awkwardness
looks like we can be fooled pretty badly about
the initial delta-v vector of a maneuver node if
the node is a long way away.

also looks like the "sqrt" I added to make the
throttle do a linear ramp, has the side effect of
not just increasing the throttle for small burns to
finish them quickly, but actually overcooking them,
so ditch the sqrt and return to linear.