mnv:step stop when burnvector turns
back to CheersKevin termination condition:
stash the burn vector and stop when current burnvector
is more than 90 degrees away from it. currenlty using vang,
could use dot product if 90 degrees is going to be locked.

observed that as we approach a node from far away, the
burn vector rotates. can be observed: when far, align to the
burn vector, then warp. watch it drift away.

so save the burn vector every iteration, until we pass the
scheduled time of the burn.