Brace adjust for VS Code "folding"
VS-Code "folding" looks best if the closing brace for a block
is at the end of the last line (so it folds the whole block into
just showing the first line).

If the brace is first on next line (after whitespace), the folding
does not fold it,  making blocks fold to two lines when they
could be just one.

This change pulls a bunch of braces up onto the prior line.