add match:plan_incl
set up a maneuver node at the next ascending or descending
node (targ orbit versus ship orbit) to transition ship onto the
new orbital plane.

note that this always places the node, and it might in fact
be a trivial amount of delta-v. it is up to the mission to make
the call on whether to use it or remove it.