add phase:ascent_v2 (throttle to control apoapsis eta)
Add an alternate ascent profile which modulates the throttle to keep
the ETA of the Apoapsis within a time range, and include the results
of my first experiment.

My conclusion at this point is that throttling back during ascent does
not necessarily save fuel (and in this case, it was slightly less fuel
efficient in the end), and it takes longer -- but it does indicate that
a shift could be made to engines with lower thrust, if that gives us
more efficient fuel usage.