new phase:autostager
use SHIP:MAXTHRUSTAT(0) to rapidly detect whether or not staging
is needed. Only check the engine list if maxthrust is zero.

keep the old autostager in autostager_enginelist; missions that
might boot into a condition where we have available thrust but
should stage (that is, a partially burned out asparagus) will
want to call the autostager_enginelist once during boot if they
might be in that condition.

it might also be a good idea to use the enginelist variant if
launching where the bottom stage is engines, and the next stage
is launch clamps. Better is to just keep the clamps in the bottom
stage since KSP engines instantly provide their thrust.