add phase:ap_pe
New step function: adjust current velocity vector to make the
periapsis and apoapsis approach the given values. Normally
set up in the mission main sequence list using "bind" to supply
the assigned apoapsis and periapsis altitudes.

Normally, parameters are AP and PE, but they are actually
accepted in either order.

Normally, we expect PE <= altitude <= AP. If the altitude is
not between PE and AP, this code will try to enter an orbit
with the correct semi-major axis length (and period), with the
current altitude being either PE or AP.

very similar to phase:circ. It would be possible to write
phase:circ as a call to phase:ap_pe but not efficient.