fix "stage after reboot" issue
Sometimes, when I reboot while in orbit, the autostager will fire
even though I think it should not. Add some debugging. I think I
found what was going on, and refusing to stage in the first
two calls avoids the transient condition that I think is happening.

Staging does not happen every time, and I do not have a save that
triggered it on loading, so if the cause is not what I think I found
then this will not fix the problem ... :(